place amplica headline in own file
