+++ DSNP-Wallet/Source/Modules/Onboarding/NewUI/CreateIdentity/ConfirmHandleView.swift
@@ -7,33 +7,6 @@
 
 import SwiftUI
 
-struct AmplicaHeadline: View {
-
-    var withBackButton: Bool = false
-    var backAction: () -> ()?
-    
-    private let indentationWidth: CGFloat = 18.0
-    
-    var body: some View {
-        HStack(alignment: .top) {
-            if withBackButton {
-                BackButton {
-                    backAction()
-                }
-                .padding(.leading, indentationWidth)
-                .padding(.top, 12)
-                Spacer()
-                AmplicaLogo()
-                    .padding(.trailing, indentationWidth)
-                Spacer()
-                EmptyView()
-            } else {
-                AmplicaLogo()
-            }
-        }
-    }
-}
-
 struct ConfirmHandleView: View {
     @ObservedObject var viewModel: ConfirmHandleViewModel
